Refactor routes into dedicated module
